# openapi2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WattTime Data API",
    "version": "V3",
    "x-logo": {
      "url": "https://www.watttime.org/api-documentation/images/logo-27870101.png",
      "backgroundColor": "#183B66",
      "altText": "WattTime Logo"
    }
  },
  "servers": [
    {
      "url": "https://api.watttime.org",
      "description": "WattTime Base API"
    }
  ],
...
  "paths": {
    "/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register New User",
        "description": "Provide basic information to self register for an account.",
        "operationId": "post_username_register_post",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "name of user that will be used in subsequent calls",
              "title": "Username"
            },
            "description": "name of user that will be used in subsequent calls",
            "example": "freddo"
          },
          {
            "name": "password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "user password. Password must be at least 8 characters, with at least 1 of each alpha, number and special characters.",
              "title": "Password"
            },
            "description": "user password. Password must be at least 8 characters, with at least 1 of each alpha, number and special characters.",
            "example": "the_frog"
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email",
              "description": "valid email address. The email address used to register will only be used for communication regarding API outages and updates. The email address will not be shared or used for any other promotional purpose. For others in your organization who would like these updates, they can subscribe to our Status Page. ",
              "title": "Email"
            },
            "description": "valid email address. The email address used to register will only be used for communication regarding API outages and updates. The email address will not be shared or used for any other promotional purpose. For others in your organization who would like these updates, they can subscribe to our Status Page. ",
            "example": "freddo@frog.org"
          },
          {
            "name": "org",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "organization name",
              "title": "Org"
            },
            "description": "organization name",
            "example": "freds world"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "Python",
            "source": "\n# To register, use the code below. Please note that for these code examples we are using filler values for username\n# (freddo), password (the_frog), email (freddo@frog.org), org (freds world) and you should replace each if you are\n# copying and pasting this code.\n\nimport requests\nregister_url = 'https://api.watttime.org/register'\nparams = {'username': 'freddo',\n         'password': 'the_frog',\n         'email': 'freddo@frog.org',\n         'org': 'freds world'}\nrsp = requests.post(register_url, json=params)\nprint(rsp.text)\n",
            "label": "Python"
          }
        ]
      }
    },
...
    "/password": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Password Reset",
        "description": "Provide your username to request an email be sent to you with password reset instructions.",
        "operationId": "get_password_password_get",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "name of user that will be used in subsequent calls",
              "title": "Username"
            },
            "description": "name of user that will be used in subsequent calls",
            "example": "freddo"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "Python",
            "source": "\n# To reset your password, use this code:\n\nimport requests\npassword_url = 'https://api.watttime.org/password/?username=freddo'\nrsp = requests.get(password_url)\nprint(rsp.json())\n",
            "label": "Python"
          }
        ]
      }
    },
    "/v3/region-from-loc": {
      "get": {
        "tags": [
          "Regions and Maps"
        ],
        "summary": "Determine Grid Region",
        "description": "Emissions intensity varies by location, specifically the location where an energy-using device is interconnected to the grid. This endpoint, provided with latitude and longitude parameters, returns the details of the grid region serving that location, if known, or a Coordinates not found error if the point lies outside of known/covered regions.",
        "operationId": "get_reg_loc_v3_region_from_loc_get",
        "parameters": [
          {
            "name": "signal_type",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "co2_moer",
                "co2_aoer",
                "health_damage"
              ],
              "type": "string",
              "description": "signal_type for which to look up region",
              "title": "Signal Type"
            },
            "description": "signal_type for which to look up region",
            "example": "co2_moer"
          },
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Latitude of desired location",
              "title": "Latitude"
            },
            "description": "Latitude of desired location",
            "example": -72.519
          },
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Longitude of desired location",
              "title": "Longitude"
            },
            "description": "Longitude of desired location",
            "example": 42.372
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionLocResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "Python",
            "source": "\n# Make sure to replace the parameters username (e.g. ‘freddo’) and password (e.g. ‘the_frog’) with your registered\n# credentials when using this code. You should not add in your token here. The code automatically generates a new token\n# each time you run it.\n\n\nimport requests\nfrom requests.auth import HTTPBasicAuth\n\nlogin_url = 'https://api.watttime.org/login'\ntoken = requests.get(login_url, auth=HTTPBasicAuth(‘freddo’, ‘the_frog’)).json()['token']\n\nregion_url = 'https://api.watttime.org/v3/region-from-loc'\nheaders = {'Authorization': 'Bearer {}'.format(token)}\nparams = {'latitude': '42.372', 'longitude': '-72.519', 'signal_type': 'co2_moer'}\nrsp=requests.get(region_url, headers=headers, params=params)\nprint(rsp.text)\n",
            "label": "Python"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PasswordResponse": {
        "properties": {
          "ok": {
            "type": "string",
            "title": "Ok"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "PasswordResponse",
        "example": {
          "ok": "Please check your email for the password reset link"
        }
      },
      "RegionLocResponse": {
        "properties": {
          "abbrev": {
            "type": "string",
            "title": "Abbrev"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "signal_type": {
            "type": "string",
            "title": "Signal Type"
          }
        },
        "type": "object",
        "required": [
          "abbrev",
          "name",
          "signal_type"
        ],
        "title": "RegionLocResponse",
        "example": {
          "abbrev": "ISONE_WCMA",
          "name": "ISONE Western/Central Massachusetts",
          "signal_type": "co2_moer"
        }
      },
      "RegisterResponse": {
        "properties": {
          "user": {
            "type": "string",
            "title": "User"
          },
          "ok": {
            "type": "string",
            "title": "Ok"
          }
        },
        "type": "object",
        "required": [
          "user",
          "ok"
        ],
        "title": "RegisterResponse",
        "example": {
          "ok": "User created",
          "user": "freddo"
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}